requirements.txt
django==1.8
djangorestframework==3.6.3
drf-extensions==0.3.1
django-dcors==0.1
django-filter==1.0.4
py-dateutil==2.2
chardet==3.0.4
xlrd==1.0.0
xlwt==1.2.0
requests==2.12.1
Pillow
lxml
MySQL-python==1.2.4b4
pytz==2017.2
apscheduler==3.5.0
django-role-permissions